feat: add asset and src model; build(deps): python
