+++ src/dbt/kipptaf/models/adp_workforce_now/sources-external.yml
@@ -28,3 +28,11 @@ sources:
           options:
             format: AVRO
             enable_logical_types: true
+      - name: src_adp_workforce_now__workers
+        external:
+          location: "{{ var('cloud_storage_uri_base') }}/adp/workforce_now/workers/*"
+          options:
+            format: AVRO
+            enable_logical_types: true
+            hive_partition_uri_prefix:
+              "{{ var('cloud_storage_uri_base') }}/adp/workforce_now/workers/"
