+++ src/teamster/kipptaf/adp/workforce_now/__init__.py
@@ -1,7 +1,21 @@
-from .api.jobs import _all as jobs
-from .api.schedules import _all as schedules
-from .sftp.assets import _all as assets
-from .sftp.sensors import _all as sensors
+from . import api, sftp
+
+assets = [
+    *api.assets,
+    *sftp.assets,
+]
+
+jobs = [
+    *api.jobs,
+]
+
+schedules = [
+    *api.schedules,
+]
+
+sensors = [
+    *sftp.sensors,
+]
 
 _all = [
     assets,
