+++ src/teamster/kipptaf/adp/workforce_now/api/__init__.py
@@ -1,0 +1,9 @@
+from .assets import _all as assets
+from .jobs import _all as jobs
+from .schedules import _all as schedules
+
+_all = [
+    assets,
+    jobs,
+    schedules,
+]
