+++ src/teamster/kipptaf/adp/workforce_now/sftp/__init__.py
@@ -1,0 +1,7 @@
+from .assets import _all as assets
+from .sensors import _all as sensors
+
+_all = [
+    assets,
+    sensors,
+]
