--- tests/assets/test_google_sheets_assets.py
@@ -1,1 +1,0 @@
-# TODO: tests for dagster-external assets
